Modified the writer to write to javascript format.
