+++ src/main/resources/SummarizedOuput.json
@@ -1,0 +1,70 @@
+var inputData = 
+[
+  {
+    "classes": [
+      {
+        "fileName": "Warning.java",
+        "filePath": "C:\\Users\\Clinton\\Documents\\GitHub\\ContextProject-TSE\\src\\main\\java\\BlueTurtle\\warnings\\Warning.java",
+        "warningList": [
+          {
+            "line": 1,
+            "message": "Name BlueTurtle.warnings must match pattern ^[a-z]+(\\.[a-zA-Z_][a-zA-Z0-9_]*)*$.",
+            "ruleName": "PackageName",
+            "fileName": "Warning.java",
+            "type": "CheckStyle",
+            "filePath": "C:\\Users\\Clinton\\Documents\\GitHub\\ContextProject-TSE\\src\\main\\java\\BlueTurtle\\warnings\\Warning.java"
+          },
+          {
+            "line": 13,
+            "message": "Variable category must be private and have accessor methods.",
+            "ruleName": "VisibilityModifier",
+            "fileName": "Warning.java",
+            "type": "CheckStyle",
+            "filePath": "C:\\Users\\Clinton\\Documents\\GitHub\\ContextProject-TSE\\src\\main\\java\\BlueTurtle\\warnings\\Warning.java"
+          },
+          {
+            "line": 14,
+            "message": "Variable filename must be private and have accessor methods.",
+            "ruleName": "VisibilityModifier",
+            "fileName": "Warning.java",
+            "type": "CheckStyle",
+            "filePath": "C:\\Users\\Clinton\\Documents\\GitHub\\ContextProject-TSE\\src\\main\\java\\BlueTurtle\\warnings\\Warning.java"
+          },
+          {
+            "line": 15,
+            "message": "Variable type must be private and have accessor methods.",
+            "ruleName": "VisibilityModifier",
+            "fileName": "Warning.java",
+            "type": "CheckStyle",
+            "filePath": "C:\\Users\\Clinton\\Documents\\GitHub\\ContextProject-TSE\\src\\main\\java\\BlueTurtle\\warnings\\Warning.java"
+          },
+          {
+            "line": 20,
+            "message": "Unused @param tag for filePath.",
+            "ruleName": "JavadocMethod",
+            "fileName": "Warning.java",
+            "type": "CheckStyle",
+            "filePath": "C:\\Users\\Clinton\\Documents\\GitHub\\ContextProject-TSE\\src\\main\\java\\BlueTurtle\\warnings\\Warning.java"
+          }
+        ],
+        "packageName": "BlueTurtle.warnings",
+        "numberOfWarnings": 5,
+        "warningTypes": [
+          "CheckStyle"
+        ],
+        "numberOfCheckStyleWarnings": 5,
+        "numberOfPMDWarnings": 0,
+        "numberOfFindBugsWarnings": 0
+      }
+    ],
+    "numberOfClasses": 1,
+    "packageName": "BlueTurtle.warnings",
+    "numberOfWarnings": 5,
+    "warningTypes": [
+      "CheckStyle"
+    ],
+    "numberOfCheckStyleWarnings": 5,
+    "numberOfPMDWarnings": 0,
+    "numberOfFindBugsWarnings": 0
+  }
+];
