+++ src/test/java/BlueTurtle/writers/JSWriterTest.java
@@ -1,44 +1,62 @@
 package BlueTurtle.writers;
 
-import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertSame;
 import static org.junit.Assert.assertTrue;
 
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
+import BlueTurtle.groupers.WarningGrouper;
 import BlueTurtle.summarizers.Summarizer;
+import BlueTurtle.warnings.CheckStyleWarning;
+import BlueTurtle.warnings.Warning;
 
 /**
- * Test class for JSWriter class.
+ * Tests for the JsonWriter class.
  * 
  * @author BlueTurtle.
  *
  */
 public class JSWriterTest {
-	
+
 	private String outputPath;
+	private List<Summarizer> summarizedWarnings;
 
 	/**
-	 * Make sure that there's not a file already written in the folder by
-	 * coincidence.
+	 * Intitialize the things that are needed.
 	 */
 	@Before
 	public void initialize() {
-		outputPath = "./resources/testExample.js";
+		outputPath = "./src/test/resources/testOutput.json";
+		HashMap<String, String> componentsInfo = new HashMap<String, String>();
+		componentsInfo.put("ExampleClass.java", "./src/test/resources/ExampleClass.txt");
+		Set<String> packagesNames = new HashSet<String>();
+		packagesNames.add("SomePackage.subpackage");
+		List<Warning> list = new ArrayList<Warning>();
+		list.add(
+				new CheckStyleWarning("./src/test/resources/ExampleClass.txt", "ExampleClass.java", 5, "test", "test"));
+		WarningGrouper wg = new WarningGrouper(componentsInfo, packagesNames, list);
+		summarizedWarnings = wg.groupBy("packages");
+
+		// make sure that the file does not already exist by coincidence.
 		File file = new File(outputPath);
 		if (file.exists()) {
 			file.delete();
 		}
-		
+
 	}
-	
+
 	/**
-	 * Clean up the files that were created for testing.
+	 * Cleanup the files that were created.
 	 */
 	@After
 	public void cleanup() {
@@ -47,35 +65,30 @@ public class JSWriterTest {
 	}
 
 	/**
-	 * Test that a file is written.
+	 * Test writing summarized warnings to a file.
 	 * 
 	 * @throws IOException
-	 *             throws an exception if problem is encountered while writing.
+	 *             throws an exception if something has happened when writing to
+	 *             a file.
 	 */
 	@Test
-	public void testFileIsWritten() throws IOException {
-		JSWriter writer = new JSWriter(new ArrayList<Summarizer>());
-		writer.writePackageToJS(outputPath);
+	public void testWriteToJsonFormat() throws IOException {
+		JSWriter jwriter = new JSWriter(summarizedWarnings);
+		jwriter.writeToJSFormat(outputPath);
+
 		File file = new File(outputPath);
-		assertTrue(file.exists());
+		boolean fileWritten = file.exists();
+		assertTrue(fileWritten);
 	}
 	
 	/**
-	 * Test the checkForComma method that should return true.
+	 * Test changing the summarizedWarnings.
 	 */
 	@Test
-	public void testCheckForCommaTrue() {
-		JSWriter writer = new JSWriter(new ArrayList<Summarizer>());
-		assertTrue(writer.checkCommaNeeded(3, 4));
-	}
-	
-	/**
-	 * Test the checkForComma method that should return false.
-	 */
-	@Test
-	public void testCheckForCommaFalse() {
-		JSWriter writer = new JSWriter(new ArrayList<Summarizer>());
-		assertFalse(writer.checkCommaNeeded(3, 3));
+	public void testChangingSummarizedWarnings() {
+		JSWriter jwriter = new JSWriter(summarizedWarnings);
+		jwriter.setSummarizedWarnings(new ArrayList<Summarizer>());
+		assertSame(0, jwriter.getSummarizedWarnings().size());
 	}
 
 }
